+++ zcash_client_sqlite/src/lib.rs
@@ -57,6 +57,7 @@ use zcash_client_backend::{
         self,
         chain::{BlockSource, ChainState, CommitmentTreeRoot},
         scanning::{ScanPriority, ScanRange},
+        wallet::ConfirmationsPolicy,
         Account, AccountBirthday, AccountMeta, AccountPurpose, AccountSource, AddressInfo,
         BlockMetadata, DecryptedTransaction, InputSource, NoteFilter, NullifierQuery, ScannedBlock,
         SeedRelevance, SentTransaction, SpendableNotes, TargetValue, TransactionDataRequest,
@@ -574,7 +575,8 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
         account: Self::AccountId,
         target_value: TargetValue,
         sources: &[ShieldedProtocol],
-        anchor_height: BlockHeight,
+        target_height: BlockHeight,
+        confirmations_policy: ConfirmationsPolicy,
         exclude: &[Self::NoteRef],
     ) -> Result<SpendableNotes<Self::NoteRef>, Self::Error> {
         Ok(SpendableNotes::new(
@@ -584,7 +586,8 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                     &self.params,
                     account,
                     target_value,
-                    anchor_height,
+                    target_height,
+                    confirmations_policy,
                     exclude,
                 )?
             } else {
@@ -597,7 +600,8 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                     &self.params,
                     account,
                     target_value,
-                    anchor_height,
+                    target_height,
+                    confirmations_policy,
                     exclude,
                 )?
             } else {